+++ BudgetAppCross.Models/DateRangeGrouping.cs
@@ -6,9 +6,9 @@ namespace BudgetAppCross.Models
 {
     public class DateRangeGrouping<TKey, T> : Grouping<TKey, T>
     {
-        public double Sum { get; set; }
-        public double StartingBalance { get; set; }
-        public double EndingBalance => StartingBalance - Sum;
+        public decimal Sum { get; set; }
+        public decimal StartingBalance { get; set; }
+        public decimal EndingBalance => StartingBalance - Sum;
         public BankAccount BankAccount { get; set; }
         public DateRangeGrouping(TKey key, IEnumerable<T> items) : base(key, items)
         {
